This is not a KnownBug anymore \o/
